requirements.txt
requests==2.28.2
PyQt5==5.15.7
openai==0.28
python-dotenv==0.21.0
matplotlib
